--- apecs/src/Apecs/Experimental.hs
@@ -1,2 +1,0 @@
-module Apecs.Experimental where
-
