--- apecs/src/Apecs/Experimental/System.hs
@@ -1,3 +1,0 @@
-module Apecs.Experimental.System where
-
-
